feat(toolkit)!: remove recall + fetch entirely
The local RAG system (recall: LanceDB + SQLite FTS + MLX embeddings +
launchd watcher daemon) wasn't earning its keep, and fetch existed to
feed it. Remove both end to end:

- toolkit: delete src/lib/recall, src/daemon (recall-only), src/lib/fetch,
  src/commands/fetch, both handlers, the recall skill, and their tests;
  drop @lancedb/lancedb + gray-matter; strip install.sh of the MLX pip
  install, ~/.recall dirs, skill copy, and launchd plist install
- dotfiles: delete the toolkit-recall LaunchAgent plist; replace the
  global toolkit recall/fetch instructions with direct lightpanda /
  PinchTab / Docling guidance
- temporal: drop the `toolkit recall search` semantic layer from the PR
  review pipeline — hybrid-retrieval.ts becomes lexical-only
  symbol-retrieval.ts (score = distinct-identifier hit count, RRF gone)
- docs: remove the Fetch & Recall section from AGENTS.md, update the
  toolkit descriptions in the structure tree, greptile map, and the
  monorepo architecture doc
