+++ AGENTS.md
@@ -40,7 +40,7 @@ packages/
 ├── tasks-for-obsidian/         # React Native task app
 ├── temporal/                   # Temporal workflows, schedules, and agent-task scheduler
 ├── terraform-provider-asuswrt/ # Terraform provider for AsusWRT
-├── toolkit/                    # CLI developer tools (fetch, recall, pr, pd, bugsink, grafana)
+├── toolkit/                    # CLI developer tools (pr, pd, bugsink, grafana)
 ├── trmnl-dashboard/            # TRMNL e-ink dashboard
 ├── webring/                    # Webring component (npm)
 scripts/                        # Repo automation (setup-free): checks, deploys, release, hooks
@@ -302,51 +302,9 @@ Each package has its own AGENTS.md with specific instructions:
 - `packages/homelab/AGENTS.md` - K8s, cdk8s, OpenTofu infrastructure
 - `packages/scout-for-lol/AGENTS.md` - Match analysis pipeline
 - `packages/resume/AGENTS.md` - Resume site
-- `packages/toolkit/AGENTS.md` - CLI developer tools (fetch, recall, pr, pd, bugsink, grafana)
+- `packages/toolkit/AGENTS.md` - CLI developer tools (pr, pd, bugsink, grafana)
 - `packages/tasks-for-obsidian/AGENTS.md` - React Native task app
 - `packages/docs/` - AI-maintained docs (see `monorepo-docs` skill)
-
-## Toolkit — Fetch & Recall
-
-The `toolkit` CLI (`~/.local/bin/toolkit`) provides web fetching and local RAG search across all Claude history.
-
-### Searching past work
-
-Use `toolkit recall search` to find context from previous conversations, plans, research, docs, and fetched pages **before** answering knowledge questions or starting tasks that may have been discussed before.
-
-```bash
-# Search across everything (hybrid semantic + keyword)
-toolkit recall search "how does the CI pipeline work"
-toolkit recall search "kubernetes resource limits"
-toolkit recall search "interview prep system design"
-```
-
-This searches:
-
-- All past Claude conversations (`.jsonl` files)
-- Claude plans and research (`~/.claude/plans/`, `~/.claude-extra/research/`)
-- Claude memories (`~/.claude/projects/*/memory/`)
-- Monorepo docs (`packages/docs/`)
-- Fetched web pages (`~/.recall/fetched/`)
-
-### Fetching web pages
-
-```bash
-toolkit fetch <url>              # Fetch via lightpanda, save + index
-toolkit fetch <url> --browser    # Use PinchTab CLI (real Chrome) if lightpanda is blocked
-toolkit fetch <url> --crawl      # Crawl a docs site
-```
-
-**Use `toolkit fetch` instead of raw `lightpanda fetch`** — it saves the output and auto-indexes it for future search.
-
-### Index management
-
-```bash
-toolkit recall add <path>        # Index a file or directory
-toolkit recall reindex           # Re-scan all watched directories
-toolkit recall status            # Index stats, daemon health
-toolkit recall debug             # Full diagnostic check
-```
 
 ## PR Media & Demo Artifacts — `public.sjer.red`
