+++ packages/dotfiles/AGENTS.md
@@ -7,8 +7,7 @@ Before doing ANY work, scan the available skills list for relevant skills and LO
 - **ALWAYS load matching skills first** - If the task involves a technology that has a skill (Docker, Terraform, Kubernetes, Git, TypeScript, etc.), load that skill BEFORE taking any action. Do not attempt to solve the problem without the skill.
 - **Use MCP tools first** - When MCP servers provide relevant functionality, prefer them over manual approaches
 - **Leverage plugins** - Check available plugins before implementing something from scratch
-- **Use `toolkit recall search` for knowledge lookups** - Before answering questions about past work, plans, research, or how things were done before, search the local RAG index: `toolkit recall search "<query>"`. This searches all past Claude conversations, plans, research docs, monorepo docs, and fetched web pages.
-- **Use `toolkit fetch` for web pages** - When fetching web pages or docs, prefer `toolkit fetch <url>` — it uses lightpanda under the hood, saves the content to `~/.recall/fetched/`, and auto-indexes it for future `toolkit recall search`. Use `toolkit fetch <url> --browser` for sites that block lightpanda. Fall back to raw `lightpanda fetch --dump markdown --strip_mode full --log_level fatal <url>` only if toolkit is not installed. Fall back to PinchTab CLI only for interactive page manipulation (clicking, form filling, screenshots). Load the `pinchtab-helper` skill before using PinchTab.
+- **Web access — lightpanda, PinchTab, Docling, or similar** - For plain page/docs fetches, use [lightpanda](https://github.com/lightpanda-io/browser): `lightpanda fetch --dump markdown --strip_mode full --log_level fatal <url>` (see the `lightpanda-browser` skill). For sites that block lightpanda or need interaction (clicking, form filling, screenshots), use [PinchTab](https://github.com/pinchtab/pinchtab) — load the `pinchtab-helper` skill first. For document extraction (PDF/DOCX/etc.), use [Docling](https://github.com/docling-project/docling) per the PDF Extraction section below. Other similar tools are fine when they fit better.
 - **Look deeper** - If CI is failing, a build tool is erroring, or infrastructure has issues, don't just report the surface error. Load the relevant skill and investigate the root cause. The user wants solutions, not descriptions of problems.
 
 Examples of what NOT to do:
