--- packages/dotfiles/Library/LaunchAgents/com.shepherdjerred.toolkit-recall.plist
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<!DOCTYPE plist PUBLIC "-//Apple//DTD PLIST 1.0//EN" "http://www.apple.com/DTDs/PropertyList-1.0.dtd">
-<plist version="1.0">
-<dict>
-    <key>Label</key>
-    <string>com.shepherdjerred.toolkit-recall</string>
-    <key>ProgramArguments</key>
-    <array>
-        <string>/Users/jerred/.local/bin/toolkit</string>
-        <string>recall</string>
-        <string>watch</string>
-    </array>
-    <key>RunAtLoad</key>
-    <true/>
-    <key>KeepAlive</key>
-    <true/>
-    <key>StandardOutPath</key>
-    <string>/Users/jerred/.recall/logs/daemon-stdout.log</string>
-    <key>StandardErrorPath</key>
-    <string>/Users/jerred/.recall/logs/daemon-stderr.log</string>
-    <key>WorkingDirectory</key>
-    <string>/Users/jerred</string>
-    <key>EnvironmentVariables</key>
-    <dict>
-        <key>PATH</key>
-        <string>/Users/jerred/.local/bin:/opt/homebrew/bin:/usr/local/bin:/usr/bin:/bin</string>
-    </dict>
-    <key>ThrottleInterval</key>
-    <integer>5</integer>
-</dict>
-</plist>
