+++ packages/temporal/src/lib/symbol-retrieval.test.ts
@@ -1,12 +1,10 @@
 import { describe, expect, it } from "bun:test";
 import {
-  buildRecallQueryFromDiff,
   extractIdentifiersFromDiff,
   formatRetrievedSymbols,
-  hybridSearch,
   lexicalRetrieve,
-  type RecallSearchFn,
-} from "./hybrid-retrieval.ts";
+  retrieveSymbols,
+} from "./symbol-retrieval.ts";
 import type { SymbolEntry, SymbolIndex } from "./symbol-index.ts";
 
 function makeEntry(
@@ -86,23 +84,21 @@ describe("lexicalRetrieve", () => {
     const hits = lexicalRetrieve(diff, index);
     expect(hits).toHaveLength(1);
     expect(hits[0]?.entry.name).toBe("renamedFunction");
-    expect(hits[0]?.rank).toBe(1);
+    expect(hits[0]?.score).toBe(1);
   });
 
-  it("ranks more-mentioned symbols higher", () => {
-    const index = buildIndex([
-      makeEntry("frequentlyUsed", "a.ts", 1),
-      makeEntry("rarelyUsed", "b.ts", 1),
-    ]);
-    const diff = `+frequentlyUsed(); frequentlyUsed(); frequentlyUsed();
-+const x = rarelyUsed();`;
+  it("scores symbols hit by more distinct identifiers higher", () => {
+    // Register the same entry under two names (an alias) so two distinct
+    // diff tokens resolve to it — that's what pushes count above 1.
+    const shared = makeEntry("handleThing", "a.ts", 1);
+    const index = buildIndex([shared, makeEntry("minorFn", "b.ts", 1)]);
+    index.byName.set("thingHandler", [shared]);
+    const diff = `+handleThing(); thingHandler(); minorFn();`;
     const hits = lexicalRetrieve(diff, index);
-    // Both names appear once each as deduplicated tokens; the lookup
-    // counts *distinct token → entry* hits, not raw occurrences. Both
-    // get count=1. Verify we at least surface both.
-    const names = new Set(hits.map((h) => h.entry.name));
-    expect(names.has("frequentlyUsed")).toBe(true);
-    expect(names.has("rarelyUsed")).toBe(true);
+    expect(hits[0]?.entry.name).toBe("handleThing");
+    expect(hits[0]?.score).toBe(2);
+    expect(hits[1]?.entry.name).toBe("minorFn");
+    expect(hits[1]?.score).toBe(1);
   });
 
   it("returns empty when no diff identifier matches the index", () => {
@@ -112,81 +108,21 @@ describe("lexicalRetrieve", () => {
   });
 });
 
-describe("buildRecallQueryFromDiff", () => {
-  it("prefers mixed-case (likely-symbol) tokens over lower-case tokens", () => {
-    const diff = `+const result = ScoutDataDragonClient.fetchManifest(rawString);`;
-    const query = buildRecallQueryFromDiff(diff);
-    // Mixed-case tokens like ScoutDataDragonClient and fetchManifest should
-    // come before lower-case `result`/`rawString`.
-    const tokens = query.split(" ");
-    const scoutIdx = tokens.indexOf("ScoutDataDragonClient");
-    const fetchIdx = tokens.indexOf("fetchManifest");
-    expect(scoutIdx).toBeGreaterThanOrEqual(0);
-    expect(fetchIdx).toBeGreaterThanOrEqual(0);
-  });
-
-  it("caps at 5 tokens", () => {
-    const diff = `+oneTwo(); threeFour(); fiveSix(); sevenEight(); nineTen(); elevenTwelve();`;
-    const query = buildRecallQueryFromDiff(diff);
-    expect(query.split(" ").length).toBeLessThanOrEqual(5);
-  });
-});
-
-// Module-scope stubs so unicorn/consistent-function-scoping doesn't complain
-// about hoisting the closures inside each `it` body.
-const emptyRecall: RecallSearchFn = () => Promise.resolve([]);
-
-const crossHitRecall: RecallSearchFn = () =>
-  Promise.resolve([
-    {
-      path: "/repo/pkg/a/src/a.ts",
-      title: "a",
-      chunk: "function crossHit() {}",
-      score: 0.9,
-      source: "code",
-      chunkIndex: 0,
-    },
-  ]);
-
-const outOfRepoRecall: RecallSearchFn = () =>
-  Promise.resolve([
-    // outside repo root — discarded
-    {
-      path: "/Users/jerred/.recall/fetched/somewhere.md",
-      title: "x",
-      chunk: "x",
-      score: 1,
-      source: "fetched",
-      chunkIndex: 0,
-    },
-    // in repo but file not indexed — discarded
-    {
-      path: "/repo/packages/zeta/AGENTS.md",
-      title: "x",
-      chunk: "x",
-      score: 1,
-      source: "doc",
-      chunkIndex: 0,
-    },
-  ]);
-
-describe("hybridSearch", () => {
-  it("returns top-1 by default (RARe top-1 design)", async () => {
+describe("retrieveSymbols", () => {
+  it("returns top-1 by default (RARe top-1 design)", () => {
     const index = buildIndex([
       makeEntry("targetFn", "pkg/a/src/a.ts", 10),
       makeEntry("otherFn", "pkg/b/src/b.ts", 20),
     ]);
-    const out = await hybridSearch({
+    const out = retrieveSymbols({
       diff: `+targetFn(); otherFn();`,
       index,
-      repoRoot: "/repo",
-      recallSearch: emptyRecall,
     });
     // Default k=1 — only one result even though two could match.
     expect(out).toHaveLength(1);
   });
 
-  it("returns top-k when k overridden", async () => {
+  it("returns top-k when k overridden", () => {
     // Tokens shorter than MIN_TOKEN_LENGTH (3) are filtered out of the diff,
     // so use real-looking identifier names here.
     const index = buildIndex([
@@ -194,69 +130,19 @@ describe("hybridSearch", () => {
       makeEntry("secondSym", "b.ts", 1),
       makeEntry("thirdSym", "c.ts", 1),
     ]);
-    const out = await hybridSearch({
+    const out = retrieveSymbols({
       diff: `+firstSym(); secondSym(); thirdSym();`,
       index,
-      repoRoot: "/repo",
-      recallSearch: emptyRecall,
       k: 3,
     });
     expect(out).toHaveLength(3);
   });
 
-  it("fuses lexical + semantic hits, scoring symbols seen in both higher", async () => {
-    const target = makeEntry("crossHit", "pkg/a/src/a.ts", 5);
-    const lexOnly = makeEntry("lexOnly", "pkg/b/src/b.ts", 10);
-    const index = buildIndex([target, lexOnly]);
-
-    // Both identifiers appear in the diff (lexical hits both). Recall
-    // also returns crossHit's file (semantic hit only for crossHit). RRF
-    // should push crossHit ahead since it's hit by both runs.
-    const out = await hybridSearch({
-      diff: `+crossHit();\n+lexOnly();`,
-      index,
-      repoRoot: "/repo",
-      recallSearch: crossHitRecall,
-      k: 2,
-    });
-    expect(out[0]?.entry.name).toBe("crossHit");
-    expect(out[0]?.sources).toEqual(["fused"]);
-    expect(out[1]?.entry.name).toBe("lexOnly");
-    expect(out[1]?.sources).toEqual(["lexical"]);
-  });
-
-  it("works with semantic disabled (recallSearch=null)", async () => {
-    const index = buildIndex([makeEntry("onlyLex", "x.ts", 1)]);
-    const out = await hybridSearch({
-      diff: `+onlyLex();`,
-      index,
-      repoRoot: "/repo",
-      recallSearch: null,
-      k: 1,
-    });
-    expect(out).toHaveLength(1);
-    expect(out[0]?.sources).toEqual(["lexical"]);
-  });
-
-  it("drops recall results whose path is outside the repo or not indexed", async () => {
+  it("returns empty when no signal is available", () => {
     const index = buildIndex([makeEntry("present", "a.ts", 1)]);
-    const out = await hybridSearch({
-      diff: `+present();`,
-      index,
-      repoRoot: "/repo",
-      recallSearch: outOfRepoRecall,
-    });
-    expect(out).toHaveLength(1);
-    expect(out[0]?.sources).toEqual(["lexical"]);
-  });
-
-  it("returns empty when no signal is available", async () => {
-    const index = buildIndex([makeEntry("present", "a.ts", 1)]);
-    const out = await hybridSearch({
+    const out = retrieveSymbols({
       diff: `+nothingMatches();`,
       index,
-      repoRoot: "/repo",
-      recallSearch: emptyRecall,
     });
     expect(out).toEqual([]);
   });
@@ -291,7 +177,6 @@ describe("formatRetrievedSymbols", () => {
             endLine: 4,
           },
           score: 1,
-          sources: ["lexical"],
         },
       ],
       { repoRoot },
