+++ packages/toolkit/package.json
@@ -8,8 +8,7 @@
   },
   "imports": {
     "#lib/*": "./src/lib/*",
-    "#commands/*": "./src/commands/*",
-    "#daemon/*": "./src/daemon/*"
+    "#commands/*": "./src/commands/*"
   },
   "bin": {
     "toolkit": "./src/index.ts"
@@ -17,7 +16,7 @@
   "scripts": {
     "build": "bun build ./src/index.ts --compile --external ffmpeg-static --outfile=dist/toolkit",
     "test": "bun run test:unit",
-    "test:unit": "bun test test/lib test/recall test/fetch test/daemon test/bugsink test/s3 test/deployed",
+    "test:unit": "bun test test/lib test/bugsink test/s3 test/deployed",
     "test:integration": "bun test test-integration",
     "typecheck": "tsc --noEmit",
     "lint": "eslint ."
@@ -31,12 +30,10 @@
   },
   "dependencies": {
     "@aws-sdk/client-s3": "^3.1001.0",
-    "@lancedb/lancedb": "^0.30.0",
     "asciinema-player": "^3.15.1",
     "debug": "^4.4.3",
     "discord.js": "^14",
     "discord.js-selfbot-v13": "^3.7",
-    "gray-matter": "^4.0.3",
     "zod": "^4.4.3"
   }
 }
